Fix the configuration of the indentation rule
Before this commit the indentation rule could not be configured.
An error has been thrown in rules.js on line 66.
This commit will change the configuration of the indentation to an
array.
To test the passing of a configuration the .gherkin-lintrc file will be
changed also.
Finally the example configuration in the README will be fixed to valid
JSON.