Implement no-watch rule
Disallows using @watch tag on Features and Scenarios.

The @watch tag is commonly used during development to run just a
specific feature or scenario.  But you don't want to commit the
code with @watch tag, which is however easy to mistakenly do.
This linting rule will prevent that common mistake.