Implement use-and rule (#59)
To check that repeated steps like: When, When, Then, Then, Then
Are written using And like: When, And, Then, And, And

When editing feature files, it's easy to make the mistake
of not using And when you should (e.g. when copy-pasting
existing steps to create new ones).  This linting rule ensures
that And is used consistently.

Future improvements:

I guess one might want to either apply this rule, or the opposite
of it - never using And, but repeating the step keywords instead.
For this the rule should be configurable, something like:

    use-and: ["on", "always"/"never"]