Read all feature files just once
Previously several rules used the fileName parameter to read in
the raw file to perform checks like trailing newlines etc.
But the feature file had already been read in before parsing
it with Gherkin.

But we can't just pass the raw file contents instead of fileName
parameter as some rules make use of the file name.

So instead I'm now passing in a file object: {name, content}.

During this refactor I also discovered that only the file-reading
operation in linter.js used encoding parameter.  In other places
no encoding was passed in, which meant that instead of a string
a Buffer was returned, which then had to be turned into string by
calling .toString().  Now the file is read in once as UTF-8 string.