Rename parsedFile variable to feature
This name was confusing to me before already as it doesn't help
you understand that the result of parsing a file is an object that
represents a feature.

Now that I changed the second parameter to file, it became even
more confusing to understand what's the difference between parsedFile
and simply file.

So I felt a proper rename was in order.