Use splitting-to-lines to determine new-line-at-eof
This regex was broken - it only reported the EOF-error correctly
when the file was using \n as line separator.  In all other cases
it immediately reported success when there was \r character anywhere
in the file.

Now using a simpler way to check for missing newline at EOF - just split
the file into lines and check if the last line is empty, in which
case there is a newline at the end of file.