Split file to lines just once inside linter.js
After the previous changes all the rules that make use of file.content
only use it to split it to lines.  To eliminate duplication, moving
the line-splitting to linter.js and instead of file.content there's now
just file.lines.

If some rule in the future needs access to file contents that's not splitted
to lines, the content field can easily be added back.  For now, YAGNI.