Implement no-watch rule (#57)
Merging for @nene - adding no `no-restricted-tags` rule
* Implement no-watch rule

Disallows using @watch tag on Features and Scenarios.

The @watch tag is commonly used during development to run just a
specific feature or scenario.  But you don't want to commit the
code with @watch tag, which is however easy to mistakenly do.
This linting rule will prevent that common mistake.

* Rename no-watch rule to no-tags

* Make no-tags rule configurable

Using syntax: ["on", {"tags": ["tag1", "tag2"]}]

I would have actually liked to use ["on", ["tag1", "tag2"]],
but the current configuration system seems to only allow strings
or objects as configuration.  Perhaps though it's for the better -
objects will be more easily extensible in the future if the need
arises.

* Rename no-tags to no-restricted-tags

* Require tag names in config to have @-prefix