Issue #109
I have been creating a new rule to check if a step have some key words and I have some problem launching this rule.

It seems that in the gherkin-lint/src/config-verifier.js file, when calling the getRule method, the additionalRulesDirs (which contains the path to my own rules) is not specified.

To fix it, it's needed to add the additionalRulesDirs parameter in all the getRule method calls, which implies add the parameter to verifyRuleConfiguration and testSubconfig method.

Included fix and test to check that the avalaible config is allowed in the custom rules