replace package 'gherkin' with '@cucumber/gherkin'
Rationale:
npm WARN deprecated cucumber-messages@8.0.0: This package is now published under @cucumber/messages
npm WARN deprecated gherkin@9.0.0: This package is now published under @cucumber/gherkin

The rule no-partially-commented-tag-lines is deleted
as the new parser will remove the partial comments from taglines
and hence they pose no problem and it is not possible to see
these comments in the parsed output

Delete test that tests that multiple errors can be reported on same feature
file, but it turns out that the new parser does not report errors that
make it possible to distinguish this case as previous.
The test is not required because it is not required that multiple errors
are reported, we only need to ensure that all errors are eventually
reported, and other tests will do this.